requirements.txt
beautifulsoup4==4.11.1
fpdf==1.7.2
fpdf2==2.5.5
gspread==5.4.0
Pillow==9.2.0
Pyrogram==2.0.33
python-dotenv==0.20.0
requests==2.28.1
TgCrypto==1.2.3
